pyproject.toml
[project]
name = "flashcards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
